% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kern.R
\name{cv_krr_fast}
\alias{cv_krr_fast}
\title{Fast Cross-Validation for KRR}
\usage{
cv_krr_fast(
  x,
  y,
  weights = NULL,
  k_folds = 3,
  b_range = NULL,
  lambda_range = 10^(seq(-4, 2, 0.5))
)
}
\description{
Fast Cross-Validation for KRR
}
\keyword{internal}
