% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboost.R
\name{predict.weight_xgboost}
\alias{predict.weight_xgboost}
\title{Predictions for weight_xgboost objects}
\usage{
predict.weight_xgboost(
  object,
  newx,
  which = c("both", "unrevised", "revised"),
  ...
)
}
\arguments{
\item{object}{A fitted weight_xgboost object.}

\item{newx}{Matrix of new covariates.}

\item{which}{Character string. Which estimator to predict: "both", "unrevised", or "revised".}

\item{...}{Additional arguments (unused).}
}
\value{
A numeric vector or data frame.
}
\description{
Predictions for weight_xgboost objects
}
