% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnn.R
\name{predict.weight_dnn}
\alias{predict.weight_dnn}
\title{Predictions for weight_dnn objects}
\usage{
\method{predict}{weight_dnn}(object, newx, which = c("both", "unrevised", "revised"), ...)
}
\arguments{
\item{object}{A fitted weight_dnn object.}

\item{newx}{Matrix of new covariates.}

\item{which}{Character string. Which estimator to predict: "both", "unrevised", or "revised".}

\item{...}{Additional arguments (unused)}
}
\value{
A numeric vector or data frame of predictions.
}
\description{
Predictions for weight_dnn objects
}
