% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.R
\name{predict.weight_lasso}
\alias{predict.weight_lasso}
\title{Predictions for weight_lasso objects}
\usage{
\method{predict}{weight_lasso}(object, newx, which = c("both", "unrevised", "revised"), ...)
}
\arguments{
\item{object}{A fitted weight_lasso object.}

\item{newx}{Matrix of new covariates.}

\item{which}{Character string. Which estimator to predict: "both", "unrevised", or "revised".}

\item{...}{Additional arguments (unused).}
}
\value{
A numeric vector or data frame.
}
\description{
Predictions for weight_lasso objects
}
